--- a/src/LinearSolve.jl
+++ b/src/LinearSolve.jl
@@ -9,7 +9,7 @@ using ArrayInterface
 using Base: cache_dependencies, Bool
 using LinearAlgebra
 using LazyArrays: @~, BroadcastArray
-using SciMLBase: AbstractLinearAlgorithm, LinearAliasSpecifier
+using SciMLBase: AbstractLinearAlgorithm, LinearAliasSpecifier, LinearVerbosity, Verbosity, @SciMLMessage
 using SciMLOperators
 using SciMLOperators: AbstractSciMLOperator, IdentityOperator
 using Setfield