--- a/src/common.jl
+++ b/src/common.jl
@@ -79,7 +79,7 @@ mutable struct LinearCache{TA, Tb, Tu, Tp, Talg, Tc, Tl, Tr, Ttol, issq, S}
     abstol::Ttol
     reltol::Ttol
     maxiters::Int
-    verbose::Bool
+    verbose::LinearVerbosity
     assumptions::OperatorAssumptions{issq}
     sensealg::S
 end