--- a/src/simplelu.jl
+++ b/src/simplelu.jl
@@ -143,6 +143,6 @@ function SciMLBase.solve!(cache::LinearCache, alg::SimpleLUFactorization; kwargs
 end
 
 function init_cacheval(alg::SimpleLUFactorization, A, b, u, Pl, Pr, maxiters::Int, abstol,
-        reltol, verbose::Bool, assumptions::OperatorAssumptions)
+        reltol, verbose::LinearVerbosity, assumptions::OperatorAssumptions)
     LUSolver(convert(AbstractMatrix, A))
 end