@jClugstor 2025-07-28
Last thing seems to be figuring out  why the OrdinaryDiffEq integration test is failing.